feat: add nota fluxes of create and update

+++ src/main/java/com/example/Secretaria/dto/response/NotaResponse.java
@@ -5,6 +5,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
@@ -15,7 +17,7 @@ public class NotaResponse {
 
     private String titulo;
 
-    private Double nota;
+    private BigDecimal nota;
 
     private String disciplina;
